useRenderColumns extracted
